+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.4.0
+
 - Added checksum module with CRC32Sink to calculate CRC32.
 - Added counter module with ByteCounter and ByteLimit.
 - Enable optional features generation on docs.rs.